Move the client map tests into their own file...
and out of the local map tests.
